@@ -12,34 +12,33 @@
     "playground": "angular-playground"
   },
   "dependencies": {
-    "@angular/animations": "~7.2.15",
-    "@angular/common": "~7.2.15",
-    "@angular/compiler": "~7.2.15",
-    "@angular/core": "~7.2.15",
-    "@angular/forms": "~7.2.15",
-    "@angular/http": "~7.2.15",
-    "@angular/platform-browser": "~7.2.15",
-    "@angular/platform-browser-dynamic": "~7.2.15",
-    "@angular/router": "~7.2.15",
+    "@angular/animations": "~8.1.3",
+    "@angular/common": "~8.1.3",
+    "@angular/compiler": "~8.1.3",
+    "@angular/core": "~8.1.3",
+    "@angular/forms": "~8.1.3",
+    "@angular/platform-browser": "~8.1.3",
+    "@angular/platform-browser-dynamic": "~8.1.3",
+    "@angular/router": "~8.1.3",
     "core-js": "^2.5.4",
     "rxjs": "~6.5.2",
     "tslib": "^1.9.0",
-    "zone.js": "~0.8.26",
-    "express": "^4.16.4",
-    "body-parser": "^1.18.3"
+    "zone.js": "~0.9.1",
+    "express": "^4.17.1",
+    "body-parser": "^1.19.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.10.0",
-    "@angular/cli": "~7.3.9",
-    "@angular/compiler-cli": "~7.2.15",
-    "@angular/language-service": "~7.2.15",
+    "@angular-devkit/build-angular": "~0.801.2",
+    "@angular/cli": "~8.1.2",
+    "@angular/compiler-cli": "~8.1.3",
+    "@angular/language-service": "~8.1.3",
     "@types/node": "~8.9.4",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
-    "codelyzer": "~4.5.0",
+    "codelyzer": "^5.0.1",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~3.0.0",
+    "karma": "~4.2.0",
     "karma-chrome-launcher": "~2.2.0",
     "karma-coverage-istanbul-reporter": "~2.0.1",
     "karma-jasmine": "~1.1.2",
@@ -47,8 +46,8 @@
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
     "tslint": "~5.11.0",
-    "typescript": "~3.2.4",
-    "@types/googlemaps": "3.30.16",
-    "angular-playground": "^5.1.5"
+    "typescript": "~3.4.5",
+    "@types/googlemaps": "3.37.0",
+    "angular-playground": "^6.0.1"
   }
 }
